Support persistent node_modules in development environment
This change will attempt to configure a local environment for either
native or Docker Compose based development.

On a first clone, any attempt to run `script/server` will trigger a
call to `script/bootstrap` which will set up a local `.env` file
and ask whether to configure this for natively or for Docker.

Subsequent calls to `script/server` should then do the right thing.

We could look subsequently at different behaviours in the update and
setup scripts depending on this choice.

When running with Docker Compose, we currently remove the node_modules
volume and any stopped or crashed containers in order to ensure that
the environment is updated properly.