Disable scroll wheel zoom on case detail page map
The map on the case detail page is almost full width, making it hard to
scroll past if scroll wheel zooming is enabled. There are +/- buttons to
control zooming, so it’s only a minor inconvenience to disable scroll
wheel zooming, but a big win for quickly moving around the page.

Fixes #51.