Simplify case filter logic.
“My cases” shows cases that are assigned to you, in any area. As before.

“All cases” (/cases) now shows all cases – regardless of assignment or
ward/area. This means “All cases” will now, by default, include cases
from outside your ward/area, and that may or may not be assigned to you,
without needing to provide the empty `?assigned=` query param.

Setting the `assigned` query param to an empty string (as in:
`/cases?assigned=`) has no effect any more, and behaves as if the query
param was not provided – whereas previously it was the only way to
request a list of cases regardless of assignment.